+++ integration/setup/common.go
@@ -143,7 +143,7 @@ func common(ctx context.Context, config Config) error {
 					},
 					{
 						Name:    "containerlinux",
-						Version: "2191.5.0",
+						Version: "2345.3.0",
 					},
 					{
 						Name:    "coredns",