+++ service/controller/resource_set.go
@@ -35,6 +35,7 @@ import (
 	"github.com/giantswarm/azure-operator/v3/service/controller/resource/instance"
 	"github.com/giantswarm/azure-operator/v3/service/controller/resource/masters"
 	"github.com/giantswarm/azure-operator/v3/service/controller/resource/namespace"
+	"github.com/giantswarm/azure-operator/v3/service/controller/resource/release"
 	"github.com/giantswarm/azure-operator/v3/service/controller/resource/resourcegroup"
 	"github.com/giantswarm/azure-operator/v3/service/controller/resource/service"
 	"github.com/giantswarm/azure-operator/v3/service/controller/resource/tenantclients"
@@ -153,6 +154,19 @@ func NewResourceSet(config ResourceSetConfig) (*controller.ResourceSet, error) {
 		}
 
 		tenantClientsResource, err = tenantclients.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
+	var releaseResource resource.Interface
+	{
+		c := release.Config{
+			G8sClient: config.K8sClient.G8sClient(),
+			Logger:    config.Logger,
+		}
+
+		releaseResource, err = release.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -396,6 +410,7 @@ func NewResourceSet(config ResourceSetConfig) (*controller.ResourceSet, error) {
 
 	resources := []resource.Interface{
 		statusResource,
+		releaseResource,
 		tenantClientsResource,
 		namespaceResource,
 		serviceResource,