+++ CHANGELOG.md
@@ -5,9 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
-
 ## [Unreleased]
+
+
+
+
+## [4.0.0] 2020-05-05
+
+### Added
+
+- Add network policy.
+
+### Changed
+
+- Replace CoreOS VM image to Flatcar with manual migration.
+- Move containerPort values from deployment to `values.yaml`.
+
+
 
 ### Changed
 
@@ -74,7 +88,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]: https://github.com/giantswarm/aws-operator/compare/v3.0.7...HEAD
+[Unreleased]: https://github.com/giantswarm/aws-operator/compare/v4.0.0...HEAD
+
+[4.0.0]: https://github.com/giantswarm/aws-operator/compare/v3.0.7...v4.0.0
 
 [3.0.7]: https://github.com/giantswarm/aws-operator/compare/v3.0.6...v3.0.7
 [3.0.6]: https://github.com/giantswarm/aws-operator/compare/v3.0.5...v3.0.6